pyproject.toml
[project]
name = "gram-resources"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "black>=25.1.0",
    "httpx>=0.28.1",
    "lxml>=5.4.0",
    "persica",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "ujson>=5.10.0",
]

[tool.uv.sources]
persica = { git = "https://github.com/luoshuijs/Persica" }
